// src/contexts/message/types.ts
export enum types {
    ADD_ERROR_MESSAGE = "SET_MESSAGE",
    ADD_WARNING_MESSAGE = "ADD_WARNING_MESSAGE",
    ADD_INFO_MESSAGE = "ADD_INFO_MESSAGE",
    SET_SUBSCRIPTION = "SET_SUBSCRIPTION",
    CLEAR_MESSAGES = "CLEAR_MESSAGES",
    REMOVE_ERROR_MESSAGE = "REMOVE_ERROR_MESSAGE",
    REMOVE_WARNING_MESSAGE = "REMOVE_WARNING_MESSAGE",
    REMOVE_INFO_MESSAGE = "REMOVE_INFO_MESSAGE"
}